um: Set secure access mode for temporary file
Replace the insecure mode 0777 with 0700 for temporary file to prohibit
other users to change the executable mapped code.

An attacker could gain access to the mapped file descriptor from the
temporary file (before it is unlinked) in a read-only mode but it should
not be accessible in write mode to avoid arbitrary code execution.

Signed-off-by: Mickaël Salaün <mic@digikod.net>